test(voice-call): split call manager tests by scenario
(cherry picked from commit 82101b152a7a0cfa1cad713287389ef65c3bd6d7)
